requirements.txt
bcrypt==3.1.7
bleach==3.3.1
boto3==1.18.60
crispy-bootstrap5==0.4
cryptocode==0.1
Django==3.2
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-debug-toolbar==3.1.1
django-extensions==3.1.0
django-filter==2.4.0
django-heroku==0.3.1
django-model-utils==4.1.1
django-modeltranslation==0.17.1
django-reversion==3.0.8
django-simple-captcha==0.5.14
django-storages==1.11.1
django-taggit==1.5.1
easy-thumbnails==2.7.1
fpdf==1.7.2
future==0.18.2
gunicorn==20.1.0
google-api-core==1.31.0
google-auth==1.33.1
google-cloud-core==1.7.1
google-cloud-storage==1.42.3
google-cloud-translate==2.0.1
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
jsonfield==3.1.0
Pillow==8.0.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
Pygments==2.3.1
PyJWT==1.7.1
python-dateutil==2.7.3
python3-openid==3.2.0
pytz==2019.3
s3transfer==0.5.0
six==1.14.0
whitenoise==5.3.0
python-decouple==3.6
taggit-selectize==2.11.0